feat: node component
