+++ packages/uilib/src/lib/plugins/type-designer/Readme.md
@@ -5,6 +5,7 @@
 - [ ] SPannung auf voltage level
 - [ ] Remove LN - so tief gehen wir nd
 - [ ] File structure. A lot of functionality is scattered and not where it should be. Expected on 20.09
+- [ ] For now its decided not to implement separate components for different node types - achieve complete agreement
 
 This plugin serves as a collection of templates available for creating a substation schema.
