+++ packages/uilib/src/lib/plugins/type-designer/canvas.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 	import { type TypeCluster } from "./types";
     import { type DataTypeTemplates, SCDQueries, UCTypeDesigner } from "@oscd-plugins/core";
-    import Bay from "./components/bay.svelte";
     import AddComponentControls from "./components/add-component-controls.svelte";
 	import { EditorEventHandler } from "./editor-events/editor-event-handler";
     import { type CreateBayEvent } from "./editor-events/event-types";
+    import Node from "./components/node.svelte";
 
 	// TODO aufraeumen in store packen
 
@@ -49,9 +49,28 @@
 
 <div class="root" class:showSidebar bind:this={htmlRoot}>
 	{#if root}
-		<Bay {typeCluster} />
+		<div class="card-container">
+			{#each Object.entries(typeCluster) as [componentName, componentArray], componentIndex (componentIndex)}
+				{#each componentArray as component (component.id)}
+					<Node
+						componentData={Object.entries(component).map(([key, value]) => ({ key, value }))}
+						componentName={componentName}
+						componentId={component.id}
+					/>
+				{/each}
+			{/each}
+		</div>
 		<AddComponentControls onClick={onCreateBay} />
 	{:else}
 		<p>No root</p>
 	{/if}
 </div>
+
+<style>
+    .card-container {
+        display: grid;
+        grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+        grid-gap: 1rem;
+        align-items: start;
+    }
+</style>
